requirements.txt
openai==1.52.0
pandas==2.1.4
python-dotenv==1.0.0
PyPDF2==3.0.1
openpyxl==3.1.2
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.29.0
asyncio-compat==0.1.2
aiohttp==3.9.1
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.22
sentence-transformers==2.2.2